rova: qcom-caf: audio: Fix pointer-to-int-cast error
device/xiaomi/rova/qcom-caf/audio/hal/audio_extn/utils.c:3484:12: error: cast to smaller integer type 'int' from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
 3484 |     return (int)key;
      |            ^~~~~~~~
1 error generated.

+++ qcom-caf/audio/hal/audio_extn/utils.c
@@ -3481,7 +3481,7 @@ size_t audio_extn_utils_get_input_buffer_size(uint32_t sample_rate,
 
 int audio_extn_utils_hash_fn(void *key)
 {
-    return (int)key;
+    return (int)(uintptr_t)key;
 }
 
 bool audio_extn_utils_hash_eq(void *key1, void *key2)
